0116 김지민 refactor: 협력사 수정/삭제 prop drilling
